+++ .angular-cli.json
@@ -25,7 +25,8 @@
       "environmentSource": "environments/environment.ts",
       "environments": {
         "dev": "environments/environment.ts",
-        "prod": "environments/environment.prod.ts"
+        "prod": "environments/environment.prod.ts",
+        "puff": "environments/environment.puff.ts"
       }
     }
   ],
