+++ src/app/app.module.ts
@@ -9,7 +9,6 @@ import { DayListComponent } from './day-list/day-list.component';
 import { EditDayComponent } from './edit-day/edit-day.component';
 import {AngularFireModule, AngularFire} from "angularfire2";
 import {RouterModule} from "@angular/router";
-import {firebaseConfig, authConfig} from "../environments/firebase.config";
 import {AuthService} from "./shared/security/auth.service";
 import {LoginComponent} from "./login/login.component";
 import {RegisterComponent} from "./register/register.component";
@@ -20,6 +19,7 @@ import {DayResolver} from "./shared/model/day.resolver";
 import {DayService} from "./shared/model/day.service";
 import { BlurChangeDirective } from './directives/blur-change.directive';
 import {TagInputModule} from "ng2-tag-input";
+import {authConfig, firebaseConfig} from "../environments/environment";
 
 @NgModule({
   declarations: [
