+++ src/environments/environment.puff.ts
@@ -10,11 +10,12 @@ export const environment = {
 };
 
 export const firebaseConfig = {
-  apiKey: "AIzaSyBVw0DXIooFx2m8oivH4aY0lA5i6HGuSKQ",
-  authDomain: "days-of-my-life-df8c0.firebaseapp.com",
-  databaseURL: "https://days-of-my-life-df8c0.firebaseio.com",
-  storageBucket: "days-of-my-life-df8c0.appspot.com",
-  messagingSenderId: "1035429128147"
+  apiKey: "AIzaSyCt7VglJnVe0ZwbFFAwXLviYCh02zEa7s0",
+  authDomain: "domlafork.firebaseapp.com",
+  databaseURL: "https://domlafork.firebaseio.com",
+  projectId: "domlafork",
+  storageBucket: "domlafork.appspot.com",
+  messagingSenderId: "805488029767"
 };
