--- src/environments/firebase.config.js
@@ -1,13 +1,0 @@
-"use strict";
-var angularfire2_1 = require("angularfire2");
-exports.firebaseConfig = {
-    apiKey: "AIzaSyBVw0DXIooFx2m8oivH4aY0lA5i6HGuSKQ",
-    authDomain: "days-of-my-life-df8c0.firebaseapp.com",
-    databaseURL: "https://days-of-my-life-df8c0.firebaseio.com",
-    storageBucket: "days-of-my-life-df8c0.appspot.com",
-    messagingSenderId: "1035429128147"
-};
-exports.authConfig = {
-    provider: angularfire2_1.AuthProviders.Password,
-    method: angularfire2_1.AuthMethods.Password
-};
